redesigned how the chunkeditor works, added object sizes
select the item you want to place, then select a position on the grid,
instead of the other way around how it worked before.
the grid now shows the size of the object, and the grid will destroy the
other object if the grid position you click on is already occupied.
the bug with overlapping nodes not deleted properly is noted